LayoutTests/media/media-source/media-source-rvfc-paused-expected.txt


EVENT(sourceopen)
EXPECTED (0 == '0') OK
EXPECTED (0 == '0') OK
EXPECTED (640 == '640') OK
EXPECTED (480 == '480') OK
EVENT(seeked)
EXPECTED (9.916667 == '9.916667') OK
EXPECTED (9.916666666666666 == '9.916667') OK
EXPECTED (640 == '640') OK
EXPECTED (480 == '480') OK
